+++ services/ui/src/components/ProjectDetailsSidebar/index.js
@@ -11,7 +11,7 @@ import { bp, color, fontSize } from 'lib/variables';
 import { Mutation } from 'react-apollo';
 
 import ProjectByNameQuery from 'lib/query/ProjectByName';
-import switchActiveStandbyMutation from 'lib/mutation/switchActiveStandby';
+import SwitchActiveStandbyMutation from 'lib/mutation/SwitchActiveStandby';
 
 const Project = ({ project }) => {
   const [copied, setCopied] = useState(false);
@@ -103,7 +103,7 @@ const Project = ({ project }) => {
               <label>Active Environment</label>
               <div className="field">
                 <Mutation 
-                  mutation={switchActiveStandbyMutation} 
+                  mutation={SwitchActiveStandbyMutation} 
                   refetchQueries={[
                     { query: ProjectByNameQuery, variables: { name: project.name } }
                   ]}
